requirements.txt
requests==2.31.0
python-dotenv==1.0.0
atlassian-python-api==3.41.0
tabulate==0.9.0
